Rename NewGameSetup.cs to MainMenu.cs
